+++ codewars_cli/consts.py
@@ -1,0 +1,20 @@
+LANGUAGE_EXTS = {
+    "python": ".py",
+    "rust": ".rs",
+}
+
+SORT_BY_OPTIONS = [
+    "newest",
+    "oldest",
+    "popularity",
+    "positive-feedback",
+    "most-completed",
+    "least-completed",
+    "recently-published",
+    "hardest",
+    "easiest",
+    "name",
+    "low-satisfaction",
+]
+
+PROGRESS_OPTIONS = ["all", "not-trained", "not-completed", "completed"]
